Integrated RabbitMQ messaging into main application

+++ Server/api/router/file.py
@@ -1,9 +1,10 @@
-from uuid import UUID
+from uuid import UUID, uuid4
+from typing import Optional
 import os
 import urllib
 
 from fastapi import APIRouter, Depends
-from fastapi.responses import StreamingResponse
+from fastapi.responses import StreamingResponse, RedirectResponse
 
 import api.service.file as file_service
 import gzip
@@ -13,28 +14,42 @@ from core.schemas.user import UserDb, User
 from core.auth.auth import get_current_user
 
 from starlette.requests import Request
-from typing import List
 from core.schemas.permission import Permission
 
-from core.exceptions import NotFoundException
+from core.middleware.redirect_middleware import redirect_middleware
 
 router = APIRouter()
 
 
 @router.post("/{file_name}", response_model=FileInDb)
-async def new_file(file_name: str, file: Request, current_user: UserDb = Depends(get_current_user)):
-    created_file = await file_service.create_file_service(file_name, file, current_user)
+async def new_file(file_name: str, file: Request, file_id: Optional[UUID] = None, current_user: UserDb = Depends(get_current_user)):
+    if not file_id:
+        file_id = uuid4()
+        redirect_host = await redirect_middleware(file_id)
+        if redirect_host:
+            host, port = redirect_host
+            new_url = file.url.replace(hostname=host, port=port).include_query_params(file_id=file_id)
+            print("New URL: ", new_url)
+            return RedirectResponse(new_url)
+
+    created_file = await file_service.create_file_service(file_name, file, current_user, file_id)
     return created_file
 
 
 @router.get("/{file_id}")
-async def download_file(file_id: UUID, current_user: UserDb = Depends(get_current_user)):
+async def download_file(file_id: UUID, req: Request, current_user: UserDb = Depends(get_current_user), redirect_host=Depends(redirect_middleware)):
+    if redirect_host:
+        host, port = redirect_host
+        new_url = req.url.replace(hostname=host, port=port)
+        return RedirectResponse(new_url)
+
     file = await file_service.get_file_by_id_service(file_id, current_user)
 
     def file_stream_generator(file_to_stream, block_size):
         full_file_path = os.path.join(file_to_stream.filepath, str(file_to_stream.id))
 
-        with gzip.open(full_file_path, 'rb') as f:
+        # with gzip.open(full_file_path, 'rb') as f:
+        with open(full_file_path, 'rb') as f:
             while True:
                 content = f.read(block_size)
                 if not content:
@@ -47,7 +62,12 @@ async def download_file(file_id: UUID, current_user: UserDb = Depends(get_curren
 
 
 @router.put("/{file_id}")
-async def update_file(file_id: UUID, file: Request, current_user: UserDb = Depends(get_current_user)):
+async def update_file(file_id: UUID, file: Request, current_user: UserDb = Depends(get_current_user), redirect_host=Depends(redirect_middleware)):
+    if redirect_host:
+        host, port = redirect_host
+        new_url = file.url.replace(hostname=host, port=port)
+        return RedirectResponse(new_url)
+
     await file_service.update_file_service(file_id, file, current_user)
 
 
@@ -67,5 +87,10 @@ async def rename_file(file_id: UUID, file_rename: File, current_user: UserDb = D
 
 
 @router.delete("/{file_id}")
-async def delete_file(file_id: UUID, current_user: UserDb = Depends(get_current_user)):
+async def delete_file(file_id: UUID, req: Request, current_user: UserDb = Depends(get_current_user), redirect_host=Depends(redirect_middleware)):
+    if redirect_host:
+        host, port = redirect_host
+        new_url = req.url.replace(hostname=host, port=port)
+        return RedirectResponse(new_url)
+
     await file_service.delete_file_by_id_service(file_id, current_user)
